Delay searching by 500 ms on each keypress
Currently it searches on every single key press (including starting a search so single letter domains so on until you reach your desired domain). So if you want to search trial.com you would consume 9 API calls (one for t, one for tr, one for tri, etc etc). That is a bit ridiculous and should be delayed for a bit so two simultaneous key presses (or very very close) don't trigger multiple API calls.

This delays the searching by 500ms on each keypress, that may be a bit high so I'm open for suggestions.